Add modules for socket events
