+++ client/js/socket-events/names.js
@@ -1,0 +1,5 @@
+"use strict";
+const socket = require("../socket");
+const render = require("../render");
+
+socket.on("names", render.renderChannelUsers);
