+++ client/js/socket-events/topic.js
@@ -1,0 +1,11 @@
+"use strict";
+const $ = require("jquery");
+const socket = require("../socket");
+const helpers_parse = require("../libs/handlebars/parse");
+
+socket.on("topic", function(data) {
+	var topic = $("#chan-" + data.chan).find(".header .topic");
+	topic.html(helpers_parse(data.topic));
+	// .attr() is safe escape-wise but consider the capabilities of the attribute
+	topic.attr("title", data.topic);
+});
